Provide a unified installer script.
The script should internally detect whether it is running on CentOS or
Ubuntu and act accordingly.

This change also includes a small enhancement: if the script is a
part of a git repository it will install yb_migrate from the local
sources. Otherwise, it will invoke `go install` with the yb_migrate
package URL.