Added new short flags -l and -p for installling from local source and only pg support to work with getopt of macos (#497)
We use the getopt utility to parse the flags passed along while running the installer script. The getopt for macOS does not have support for long options and also processes the flags in a different format. Hence I have added two new flags and the code to parse the flags depending on the macOS format or Linux format.
The two new flags are:
-l : short hand for --install-from-local-source
-p: short hand for --only-pg-support
Let me know if you want the flag names to be changed.