Introduced pg_query module for SQL parsing and updating the schema as per assessment report recommendations (#1499)
- it can handle/update existing 'WITH' clauses to add/append 'COLOCATION = true'
- also handles clauses like ON COMMIT and TABLESPACE if present after the WITH clause

Some key points about minimizing the risk/failures:

We are using this SQL parser only for DDLs where we need to apply sharding recommendations(i.e. sharded tables), the rest of the tables' DDLs are untouched.
In the worst case, if any error occurs due to this new module, we will fall back to the original DDL, and in that case, we will inform the user also about the failure to apply the recommendation for this table, and they can do the change manually.

* Informing the user in case of failure to apply a recommendation for a particular table, to manually apply it.