Use conn pool instead of raw connections in PG/YB source DBs (#1459)
Recently, a customer ran into an error `write: broken pipe` when trying out live migration (export-data). This happened because we were making some metadata queries after snapshot step, by which time, the connection that we had opened earlier was automatically closed (because of server side settings - 
```
idle_in_transaction_session_timeout = 20000
statement_timeout = 120000
idle_session_timeout = 600000
```


To avoid this, we already moved the logic of metadata query to be called before the snapshot step. (in an older PR)
To add in additional layer of safety, we are making the following changes as part of this PR:
- use the sql.DB interface instead of managing connections ourselves. This is essentially a connection pool, which is concurrency-safe.
- Additionally, we set the SetMaxOpenConns to 1, just to cap the number of connections we open to the source database. We typically never need more than one connection. (we only query some metadata). 
- We also SetConnMaxIdleTime = 5 min. This is to ensure that the pool automatically discards connections if they are idle for longer than 5 min. It is still possible for the server to have lower thresholds for idle connections; this is just some arbitrary (and hopefully conservative) value that we are setting.