Bug Fix: Do not send the call home diagnostics if startTime is not initialised (#1613)
Issue - if someone hasn't provided the correct export-dir because of which initialization step isn't run and the command fails with export-dir doesn't exist. We send payload on exit as well so in payload we were having start-time as uninitialized ("0001-01-01 00:00:00") and time taken (9223372037) was getting calculated as a very big value so the callhome was not accepting the request due to a large value in int column psycopg2.errors.NumericValueOutOfRange: value "9223372037" is out of range for type integer

Fix - Check if startTime is initialized, and based on that decide whether to send payload or not. (https://yugabyte.atlassian.net/browse/DB-12040)