FIx for unsupported types in export data from target (#1674)
Current query to fetch data types was incorrectly also checking for table names , not just data types. In one of the tests, there was a table name called "integer" which essentially cause all integer columns to be marked as unsupported.

Changed the query to specifically look for columns of input table.