Filter Issues by MinimumVersionFixedIn (#1971)
- New flag: target-db-version flag in analyze-schema/assess-migration 
user has to pass in the full version (A.B.C.D)
only versions from a list of supported series can be passed. This list will be all the YB series that are not EOL. (Although officially we document that voyager only supports the two latest stable releases and one latest preview, we would NOT like to fail here if user passes 2.18, for example.)
If user passes a version from a series that is not supported by voyager (say 2026.1.x.x), voyager will print the error, prompt the user if they want to continue with default (latest stable).
Defaults to latest stable version (2024.1.3.0, for example)

- Every Issue that we report will have the following details
For every series (from the supported list), if it is fixed in any version of the series, we store that information in "minVersionsFixedIn" which is a map[series]version
2.20: fixed in 2.20.8.0
2024.1: fixed in 2024.1.1.0
2.23: fixed in 2.23.5.1
2024.2: fixed in 2024.2.1.0

- Checking if issue is fixed in target-db-version (specified by user)
We compare target-db-version with Issue.minVersionsFixedIn[target-db-version.Series]

- Getting list of versions that issue is fixed in (to display in reports):
Option1: return all values from "minVersionsFixedIn"
Option2: return values from "minVersionsFixedIn" whose series >= target-db-version.Series. Caveat: We will not be able to compare between stable and preview releases (no way to determine if 2.23 > 2024.1 or not). Therefore for target-db-version=2024.2.1.1(stable), we will omit the older stable series (2024.1 series) but include all preview series - (2.21, 2.23).