Misc changes to target-db-version flag (#2098)
- Unhide the flag
- Print the version being used in analyze-schema/ assess-migration
- Prevent users from passing build number in target-db-version.
The problem is that 2024.2.0.0-b145 would be considered to be > 2024.2.0.0. Which means we will have two options:
a. prevent users from passing build number
b. Always ask users to pass build number, so that it can be compared against the issues' MinimumVersionsFixedIn(which should also have build number.)
Going with option a.