Make voyager import-data metadata tables columns as TEXT (#2130)
Currently, voyager metadata tables have columns of type VARCHAR(250). In one of our tests, the table name + export-dir turned out to be longer than 250, thereby leading to data_file_name field exceeding the 250 limit and therefore preventing import-data from running successfully. There's no real reason to limit any of the fields to 250.
There isn't any documented difference in performance between the VARCHAR(n) and TEXT, even in postgres

Going with ALTER TABLE ALTER COLUMN TYPE so as to ensure that the type is changed even when a user is using voyager on a cluster where voyager was previously used.

ALTER table from VARCHAR to TEXT is a safe operation. It does not lead to a table-rewrite and is an idempotent operation.