Changed installation of postgresql-17 to postgresql-client-17 in the installer script (#2323)
Issue: The postgresql-17 package on ubuntu installs the 'server' package and the postgresql-client-17 package installs the the client package which includes binaries like pg_dump and pg_restore.
If no client package is present on the system then the postgres-17 package also installs the corresponding client package. However, if a client package is already present then it is not installed/updated.
Thats why the jenkins pipeline was facing an error. The server was getting updated to 17.3 but the pg_dump was still at 17.2

Fix: Changed installation of postgresql-17 to postgresql-client-17 in the installer script