New colocated picker strategy (#2368)
Introduced a new picker:
The goal of this picker is to pick a combination of colocated and sharded tasks, both at random.
The limits in place are maxShardedTasksInProgress, maxColocatedTasksInProgress and colocatedBatchTaskQueue.

Colocated tasks are limited by single tablet performance limits on YB (by definition), so we have to constrain the no. of colocated batches that can be ingested at a time. THis is achieved by having a max parallel limit on the consumer of colocatedBatchTaskQueue.
Therefore, colocated tasks are prioritized and picked but only if colocatedBatchTaskQueue has space.
If colocatedBatchTaskQueue is full, then sharded tasks are picked.