Remove ora2pg assessment report generation in export-schema (#2405)
We had initially added this because in analyze-schema, we used to report "migration complexity".
Currently, we don't report "migration complexity" in analyze-schema, only in assess-migration.

Therefore, we don't need to generate the ora2pg assessment report in export-schema flow anymore.

Motivation for this change: There was a bug in this code path, due to which export-schema was failing.