Fix: lookup failure while creating registered list from name registry for the tables that are present on other side DB (#2407)
fixes #2406
Solution - Added a new `LookupTableNameAndIgnoreOtherSideMappingIfNotFound` name registry API for getting a name tuple with only one side object name populated (such as SourceName or TargetName) if other side mappings are not found for a table to handle the cases where a table exists in the source but not in target or vice versa.

Integration tests and unit tests are also added for the above.