Updated the pg grant permission script to prompt the user for permission before transferring ownership of tables to replication group (#2504)
Added a prompt to ask the user whether they want to proceed with the transfer of ownership of tables to the replication group
This is done just before the transfer happens in the script
The user can only provide 'yes' or 'no' as input. They cannot provide 'y' or 'n' since the psql \if does not support string comparision. 'yes' however is considered as 1 and 'no' as 0.
Issue: #2485