Fix: ANALYZE guardrail checks to consider the auto analyze on PG and gracefully error out if No is answered in prompt (#2856)
If auto-analyze is enabled on the PG, then the last_autoanalyze column is updated with the timestamp of the last analyze run automatically in the database. So, considering both the values last_analyze / last_autoanalyze to check if analyze is not run at all on the db

Erroring out gracefully in case the user answers no to the ANALYZE prompt.