Modifying btree secondary indexes to be range-sharded (#2896)
Modifying the secondary indexes using the BTREE index access method to be range-sharded indexes, as that is the only method that supports ordered scans in PG.  #2804
Fixed anonymization of object names in schema_optimization_report changes to callhome to consider qualified names.
Added unit tests in transformer_test for new transformers.