[Error Policy] Report processing errors count (#3017)
Modify the way processing errors are stored.
data/errors
└── table::public."large_row"
   └── file::large_row.csv:f0f01c9d
       └── processing-errors.1.1.1049579.log

The format is processing-errors.<batchNumber>.<rowCount>.<bytecount>.log
This facilitates two things:

processing errors are now associated with a batch (i.e. they are the errors we encountered when processing the batch). On stop/start(resume), if some errors were processed, but the final batch was not created, the process will be re-done, and the processing errors file will be rewritten, preventing duplicate error logs in the file.
storing the row count, byte count in the file name allows for import-data-status, snapshot-reports to properly calculate the no. of errors.