Added support to restore only the sequences that are attached to tables being migrated in offline migration to handle case if any of them is not present in target (#3074)
Fixes Issue #3075
In cases where the user excludes some tables in the exclude-table-list, the behavior should be that we ideally exclude the table completely. If the sequence is attached to the table, we should also skip it.

The changes in the PR for the offline migration:

For source type is PG -
We are sure that all the sequences being migrated are attached to the tables, so we can skip the sequence if it is not attached to a table in the table list
And in a scenario where the table is not present in the target database, we will skip the sequences attached to tables not part of the table list

For any other sources
No changes