Handle tables having expression based unique indexes by executing the events for them sequentially (#3189)
For the tables having this type of unique index, partitioned by table, where all the events are applied sequentially to avoid any potential issues of conflict because of parallelization
The rest of the tables will be partitioned by PK.
Added a configuration to override the CDC Partiiton strategy (PK, TABLE, and AUTO) (--cdc-partition-strategy). AUTO is the default behaviour.

For the fall-forward/fallback,
PG/ORACLE - partition strategy is partition by table, as observed that there is no huge difference in the performance between partition by pk and partition by table. So better to have a partition by table to avoid any conflict.

This PR has a breaking change, so changed the breaking release to the upcoming release