Skip multi range or custom range type columns in live migration (#3241)
For the built-in MulitRange type or user-defined range type, the export from debezium is as BYTES data, but the value converter we have for BYTES is specifically for the bytea data and doesn't work for these datatypes.#3212

For the XML datatype, it's an unsupported datatype of YB, as inserts don't work that way. We report it as unsupported completely for all workflows, and don't have handling for that in any of the workflows, and there is no way to skip this column in the snapshot part, as there is no way to skip any column via pg_dump. For that, added the Unsupported XML feature error msg in non-retryable errors to error out directly for that. #3220